requirements.txt
pytest==7.3.0
selenium == 4.0
webdriver-manager == 3.8.5
pytest-bdd == 6.1.1
allure-pytest == 2.13.1
pytest-rerunfailures == 11.1.2
